+++ src/Components/Home/HourlyForecast.tsx
@@ -1,77 +1,76 @@
 import React, { FC } from 'react';
 import styled from 'styled-components';
 
-import brokenClouds from '../../Assets/icons-sm-weather/icon-broken-clouds.svg';
-import clearSky from '../../Assets/icons-sm-weather/icon-clear-sky.svg';
-import fewClouds from '../../Assets/icons-sm-weather/icon-few-clouds.svg';
-import mist from '../../Assets/icons-sm-weather/icon-mist.svg';
-import rainyThunder from '../../Assets/icons-sm-weather/icon-rainy-thunder.svg';
-import rainy from '../../Assets/icons-sm-weather/icon-rainy.svg';
-import scatteredClouds from '../../Assets/icons-sm-weather/icon-scattered-clouds.svg';
-import showerRainy from '../../Assets/icons-sm-weather/icon-shower-rainy.svg';
-import snow from '../../Assets/icons-sm-weather/icon-snow.svg';
+import brokenClouds from 'Assets/icons-sm-weather/icon-broken-clouds.svg';
+import clearSky from 'Assets/icons-sm-weather/icon-clear-sky.svg';
+import fewClouds from 'Assets/icons-sm-weather/icon-few-clouds.svg';
+import mist from 'Assets/icons-sm-weather/icon-mist.svg';
+import rainyThunder from 'Assets/icons-sm-weather/icon-rainy-thunder.svg';
+import rainy from 'Assets/icons-sm-weather/icon-rainy.svg';
+import scatteredClouds from 'Assets/icons-sm-weather/icon-scattered-clouds.svg';
+import showerRainy from 'Assets/icons-sm-weather/icon-shower-rainy.svg';
+import snow from 'Assets/icons-sm-weather/icon-snow.svg';
 
-interface HourlyForcastProps {}
+interface HourlyForecastProps {}
 
-const HourlyForcast: FC<HourlyForcastProps> = ({}) => {
+const HourlyForecast: FC<HourlyForecastProps> = ({}) => {
   return (
-    <SHourlyForcastLayout>
-      <SforcastWrap>
+    <SHourlyForecastLayout>
+      <SforecastWrap>
         <img src={rainy} alt='' />
         <p>22&#186;</p>
         <p>오후 12시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={rainyThunder} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={snow} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={brokenClouds} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={clearSky} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={fewClouds} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={mist} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={scatteredClouds} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-      <SforcastWrap>
+      </SforecastWrap>
+      <SforecastWrap>
         <img src={showerRainy} alt='' />
         <p>26&#186;</p>
         <p>오후 3시</p>
-      </SforcastWrap>
-    </SHourlyForcastLayout>
+      </SforecastWrap>
+    </SHourlyForecastLayout>
   );
 };
 
-export default HourlyForcast;
+export default HourlyForecast;
 
-const SHourlyForcastLayout = styled.article`
+const SHourlyForecastLayout = styled.article`
   display: flex;
   align-items: center;
   gap: 20px;
-  max-width: 247px;
   padding: 11.5px 20px;
   background-color: #fff;
   box-sizing: border-box;
@@ -85,7 +84,7 @@ const SHourlyForcastLayout = styled.article`
   }
 `;
 
-const SforcastWrap = styled.div`
+const SforecastWrap = styled.div`
   text-align: center;
   flex-shrink: 0;
